+++ docs/CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.6.0] - 2024-04-20
+
+### Added
+
+- All Minecraft server commands now require user to have Minecraft role
+- Configuration (read and write) for Minecraft role in config.ini
+- Missing !help message for bubblewrap command
+
+### Changed
+
+- Updated requirements.txt to reflect all dependancies
+- Minecraft Server IP now sends via DM rather than to the main channel to avoid potential security issues
+- Stopping the Minecraft server with !stopmc now closes the tmux terminal session as well, to avoid potential for RCE
+
+### Known Issues
+
+- If the Minecraft server is stopped using !stopmc, it currently can not be restarted via Harbinger (a new tmux pane must be manually created)
+- If the Minecraft server is stopped using '!mc stop' command, the aforementioned vulnerability still exists.  This will hopefully be rectified in a future release
+
 ## [5.5.2] - 2024-04-17
 
 ### Hotfix